# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Equity Ranking and Risk API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: equity
    description: Equity ranking and screening
  - name: portfolio
    description: Portfolio construction and risk
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /equity/status:
    get:
      operationId: getDataStatus
      tags: [equity]
      summary: Get data loading status
      description: Returns current status of data loading (loading, ready, error)
      responses:
        "200":
          description: Status response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataStatus"

  /equity/rankings:
    get:
      operationId: getRankings
      tags: [equity]
      summary: Get ranked equity universe
      description: Returns ranked stocks with all factor scores, alpha, cluster labels
      parameters:
        - name: volAdjust
          in: query
          schema:
            type: boolean
          description: Apply Sharpe-style vol adjustment to momentum
        - name: useQuality
          in: query
          schema:
            type: boolean
          description: Include quality factor in alpha
        - name: useTstats
          in: query
          schema:
            type: boolean
          description: Include OLS t-stats
        - name: w6
          in: query
          schema:
            type: number
          description: Weight on 6-month momentum (default 0.4)
        - name: w12
          in: query
          schema:
            type: number
          description: Weight on 12-month momentum (default 0.4)
        - name: wQuality
          in: query
          schema:
            type: number
          description: Weight on quality factor (default 0.2)
        - name: volFloor
          in: query
          schema:
            type: number
          description: Volatility floor (default 0.05)
        - name: winsorP
          in: query
          schema:
            type: number
          description: Winsorization percentile (default 2)
        - name: clusterN
          in: query
          schema:
            type: integer
          description: Top N stocks to cluster (default 100)
        - name: clusterK
          in: query
          schema:
            type: integer
          description: Number of clusters (default 10)
        - name: clusterLookback
          in: query
          schema:
            type: integer
          description: Lookback days for correlation clustering (default 252)
        - name: secFilerOnly
          in: query
          schema:
            type: boolean
          description: Only include SEC-filing companies (default false)
        - name: excludeSectors
          in: query
          schema:
            type: string
          description: Comma-separated list of sectors to exclude
        - name: requireQuality
          in: query
          schema:
            type: boolean
          description: Only include stocks with quality coverage (default false)
      responses:
        "200":
          description: Ranked stock list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingsResponse"
        "202":
          description: Data is still loading
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataStatus"

  /equity/universe-filters:
    post:
      operationId: applyUniverseFilters
      tags: [equity]
      summary: Apply custom universe filters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UniverseFilters"
      responses:
        "200":
          description: Filtered universe rankings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingsResponse"

  /portfolio/risk:
    post:
      operationId: computePortfolioRisk
      tags: [portfolio]
      summary: Compute portfolio risk metrics
      description: Computes covariance matrix, portfolio vol, avg correlation, cluster distribution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioRiskRequest"
      responses:
        "200":
          description: Portfolio risk metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioRiskResponse"

  /portfolio/history:
    post:
      operationId: computePortfolioHistory
      tags: [portfolio]
      summary: Compute portfolio equity curve
      description: >
        Computes a weighted log-return equity curve using static current weights.
        Requires ≥252 shared valid daily return observations after alignment.
        Not a backtest — uses fixed weights applied uniformly over the lookback window.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioHistoryRequest"
      responses:
        "200":
          description: Portfolio equity curve and summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioHistoryResponse"

  /portfolio/corr-seed:
    post:
      operationId: computeCorrSeed
      tags: [portfolio]
      summary: Greedy correlation-constrained basket seeding
      description: >
        Selects up to N tickers from the candidate list such that each added ticker has
        a maximum pairwise absolute Pearson correlation ≤ maxCorr with all already-selected names.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorrSeedRequest"
      responses:
        "200":
          description: Selected basket tickers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrSeedResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DataStatus:
      type: object
      properties:
        status:
          type: string
          enum: [loading, ready, error]
        message:
          type: string
        progress:
          type: ["number", "null"]
        total:
          type: ["integer", "null"]
        loaded:
          type: ["integer", "null"]
        cachedAt:
          type: ["string", "null"]
        enrichment:
          type: string
          enum: [pending, loading, complete]
        qualityCoverage:
          type: string
        qualityEpoch:
          type: integer
        timings:
          type: object
          additionalProperties:
            type: number
      required:
        - status
        - message

    Stock:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        sector:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        marketCap:
          type: ["number", "null"]
        adv:
          type: ["number", "null"]
        m6:
          type: ["number", "null"]
        m12:
          type: ["number", "null"]
        r1:
          type: ["number", "null"]
        sigma6:
          type: ["number", "null"]
        sigma12:
          type: ["number", "null"]
        s6:
          type: ["number", "null"]
        s12:
          type: ["number", "null"]
        tstat6:
          type: ["number", "null"]
        tstat12:
          type: ["number", "null"]
        quality:
          type: ["number", "null"]
        zM6:
          type: ["number", "null"]
        zM12:
          type: ["number", "null"]
        zQuality:
          type: ["number", "null"]
        zS6:
          type: ["number", "null"]
        zS12:
          type: ["number", "null"]
        zT6:
          type: ["number", "null"]
        zT12:
          type: ["number", "null"]
        zQ:
          type: ["number", "null"]
        sSleeve:
          type: ["number", "null"]
        tSleeve:
          type: ["number", "null"]
        qSleeve:
          type: ["number", "null"]
        alpha:
          type: ["number", "null"]
        rank:
          type: ["integer", "null"]
        percentile:
          type: ["number", "null"]
        cluster:
          type: ["integer", "null"]
        roe:
          type: ["number", "null"]
        roa:
          type: ["number", "null"]
        grossMargin:
          type: ["number", "null"]
        opMargin:
          type: ["number", "null"]
        deRatio:
          type: ["number", "null"]
        zRoe:
          type: ["number", "null"]
        zRoa:
          type: ["number", "null"]
        zGross:
          type: ["number", "null"]
        zOp:
          type: ["number", "null"]
        zInvLev:
          type: ["number", "null"]
        qualityMissing:
          type: ["boolean", "null"]
        qualityMissingReason:
          type: ["string", "null"]
        hasProfitabilityBucket:
          type: ["boolean", "null"]
        hasMarginBucket:
          type: ["boolean", "null"]
        hasLeverageBucket:
          type: ["boolean", "null"]
        qualityBucketCount:
          type: ["number", "null"]
      required:
        - ticker
        - name

    UniverseAudit:
      type: object
      properties:
        preFilterCount:
          type: integer
        postFilterCount:
          type: integer
        exclusions:
          type: object
          additionalProperties:
            type: integer
        sectorBreakdown:
          type: object
          additionalProperties:
            type: integer
        qualityCoverage:
          type: string
        qualityPct:
          type: number
        activeFilters:
          type: array
          items:
            type: string

    RankingsResponse:
      type: object
      properties:
        stocks:
          type: array
          items:
            $ref: "#/components/schemas/Stock"
        total:
          type: integer
        cachedAt:
          type: ["string", "null"]
        clusterCount:
          type: integer
        audit:
          $ref: "#/components/schemas/UniverseAudit"
      required:
        - stocks
        - total
        - clusterCount

    UniverseFilters:
      type: object
      properties:
        minPrice:
          type: number
        minAdv:
          type: number
        minMarketCap:
          type: number
        volAdjust:
          type: boolean
        useQuality:
          type: boolean
        useTstats:
          type: boolean
        w6:
          type: number
        w12:
          type: number
        wQuality:
          type: number
        volFloor:
          type: number
        winsorP:
          type: number
        clusterN:
          type: integer
        clusterK:
          type: integer
        clusterLookback:
          type: integer

    PortfolioHolding:
      type: object
      properties:
        ticker:
          type: string
        weight:
          type: number
      required:
        - ticker
        - weight

    PortfolioHoldingRisk:
      type: object
      properties:
        ticker:
          type: string
        weight:
          type: number
          description: Final weight after vol-target overlay
        baseWeight:
          type: number
          description: Base weight before vol-target overlay (sums to 1)
        vol:
          type: number
          description: Annualized individual volatility
        riskContrib:
          type: number
          description: Fraction of total portfolio variance contributed by this name (sums to 1)
        cluster:
          type: ["integer", "null"]
      required:
        - ticker
        - weight
        - baseWeight
        - vol
        - riskContrib

    ClusterCount:
      type: object
      properties:
        cluster:
          type: integer
        count:
          type: integer
        weight:
          type: number
      required:
        - cluster
        - count
        - weight

    PortfolioRiskRequest:
      type: object
      properties:
        tickers:
          type: array
          items:
            type: string
          description: >
            Basket tickers (manual selection). Weights are fully automated by
            the server — the client must NOT supply per-name weights.
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioHolding"
          description: >
            DEPRECATED — weight field is ignored. Send weight=1 for all entries.
            Prefer the tickers field in future.
        lookback:
          type: integer
          description: Days of history for covariance (60, 126, or 252)
        weightingMethod:
          type: string
          enum: [equal, inverse_vol, signal_vol, risk_parity, min_var, mean_variance]
      required:
        - holdings
        - lookback
        - weightingMethod

    PortfolioRiskResponse:
      type: object
      properties:
        portfolioVol:
          type: number
          description: Final portfolio vol after vol-target overlay (= risky_sleeve² × Σ computation)
        basePortVol:
          type: number
          description: Pre-overlay portfolio vol (sqrt of w_base' Σ w_base)
        volTargetMultiplier:
          type: number
          description: Equity sleeve multiplier = min(15% / basePortVol, 1.0) — capped at 1, no leverage
        grossExposure:
          type: number
          description: Total risky-sleeve weight (= volTargetMultiplier, since base weights sum to 1)
        riskySleeve:
          type: number
          description: Risky equity sleeve total weight (same as grossExposure, explicit label)
        sgovWeight:
          type: number
          description: Residual cash / SGOV weight = max(0, 1 - riskySleeve)
        diversificationRatio:
          type: number
          description: Weighted-avg individual vol / portfolio vol (base weights). DR > 1 means diversification benefit.
        effectiveN:
          type: number
          description: Effective number of positions = 1 / sum(w_base_i^2) — Herfindahl-based
        namesCapped:
          type: array
          items:
            type: string
          description: Tickers that hit the per-name cap (Risk Parity only)
        method:
          type: string
          description: Actual method used (may differ from requested if fallback triggered)
        covModel:
          type: ["string", "null"]
          description: Covariance model label (e.g. "ewma(λ=0.94)+ridge"). Null for heuristic methods.
        fallback:
          type: ["string", "null"]
          description: Non-null when a fallback was triggered — describes what happened
        volLookback:
          type: integer
          description: Days of history used for individual vol estimation
        covLookback:
          type: integer
          description: Days of history used for covariance matrix
        avgCorrelation:
          type: number
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioHoldingRisk"
        clusterDistribution:
          type: array
          items:
            $ref: "#/components/schemas/ClusterCount"
        largestWeight:
          type: number
        numHoldings:
          type: integer
      required:
        - portfolioVol
        - basePortVol
        - volTargetMultiplier
        - grossExposure
        - riskySleeve
        - sgovWeight
        - diversificationRatio
        - effectiveN
        - namesCapped
        - method
        - volLookback
        - covLookback
        - avgCorrelation
        - holdings
        - clusterDistribution
        - largestWeight
        - numHoldings

    PortfolioHistoryRequest:
      type: object
      properties:
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioHolding"
          description: >
            Equity holdings with vol-target-scaled weights (sum = risky sleeve, not 1.0).
            Weights are used as-is — not renormalized inside the endpoint.
        lookback:
          type: integer
          description: Trading-day window for price history (default 252)
        sgovWeight:
          type: number
          description: >
            Cash/SGOV weight hint from the vol-target overlay.
            The endpoint derives cashWeight defensively from sum(equity weights) rather than
            trusting this value directly, but it signals that SGOV data should be included.
      required:
        - holdings

    PortfolioHistoryResponse:
      type: object
      properties:
        dates:
          type: array
          items:
            type: string
          description: ISO-8601 date labels — length = daysUsed + 1 (includes base date at index 0)
        nav:
          type: array
          items:
            type: number
          description: Cumulative NAV starting at 100 — length = daysUsed + 1
        drawdown:
          type: array
          items:
            type: number
          description: Drawdown from running peak (%) — length = daysUsed + 1
        totalReturn:
          type: number
          description: Total return over the window (%)
        maxDrawdown:
          type: number
          description: Maximum drawdown over the window (%)
        annualizedVol:
          type: number
          description: Annualized portfolio volatility of log-returns (%)
        numDays:
          type: integer
          description: Number of shared valid daily log-return observations used
        investedWeight:
          type: number
          description: Sum of all equity holding weights actually applied in computation
        cashWeight:
          type: number
          description: Residual cash weight = max(0, 1 - investedWeight)
        daysUsed:
          type: integer
          description: Alias for numDays — number of shared valid observations
        cashMethod:
          type: string
          description: '"sgov" when SGOV prices were available; "zero" fallback otherwise'
      required:
        - dates
        - nav
        - drawdown
        - totalReturn
        - maxDrawdown
        - annualizedVol
        - numDays
        - investedWeight
        - cashWeight
        - daysUsed
        - cashMethod

    CorrSeedRequest:
      type: object
      properties:
        tickers:
          type: array
          items:
            type: string
          description: Candidate tickers in alpha-rank order (best first)
        n:
          type: integer
          description: Max basket size
        maxCorr:
          type: number
          description: Pairwise correlation ceiling (absolute value)
        lookback:
          type: integer
          description: Trading days used for correlation estimation
      required:
        - tickers
        - n
        - maxCorr
        - lookback

    CorrSeedResponse:
      type: object
      properties:
        tickers:
          type: array
          items:
            type: string
        count:
          type: integer
        requested:
          type: integer
        maxCorr:
          type: number
        lookback:
          type: integer
        candidatesScanned:
          type: integer
      required:
        - tickers
        - count
        - requested
        - maxCorr
        - lookback
        - candidatesScanned