# lib/api-zod/src/generated/types/corrSeedRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Equity Ranking and Risk API
 * OpenAPI spec version: 0.1.0
 */

export interface CorrSeedRequest {
  /** Candidate tickers in alpha-rank order (best first) */
  tickers: string[];
  /** Max basket size */
  n: number;
  /** Pairwise correlation ceiling (absolute value) */
  maxCorr: number;
  /** Trading days used for correlation estimation */
  lookback: number;
}