# lib/api-zod/src/generated/types/corrSeedResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Equity Ranking and Risk API
 * OpenAPI spec version: 0.1.0
 */

export interface CorrSeedResponse {
  tickers: string[];
  count: number;
  requested: number;
  maxCorr: number;
  lookback: number;
  candidatesScanned: number;
}