# lib/api-zod/src/generated/types/dataStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Equity Ranking and Risk API
 * OpenAPI spec version: 0.1.0
 */
import type { DataStatusEnrichment } from "./dataStatusEnrichment";
import type { DataStatusStatus } from "./dataStatusStatus";
import type { DataStatusTimings } from "./dataStatusTimings";

export interface DataStatus {
  status: DataStatusStatus;
  message: string;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  total?: number | null;
  /** @nullable */
  loaded?: number | null;
  /** @nullable */
  cachedAt?: string | null;
  enrichment?: DataStatusEnrichment;
  qualityCoverage?: string;
  qualityEpoch?: number;
  timings?: DataStatusTimings;
}