# lib/api-zod/src/generated/types/portfolioHistoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Equity Ranking and Risk API
 * OpenAPI spec version: 0.1.0
 */
import type { PortfolioHolding } from "./portfolioHolding";

export interface PortfolioHistoryRequest {
  /** Equity holdings with vol-target-scaled weights (sum = risky sleeve, not 1.0). Weights are used as-is — not renormalized inside the endpoint.
   */
  holdings: PortfolioHolding[];
  /** Trading-day window for price history (default 252) */
  lookback?: number;
  /** Cash/SGOV weight hint from the vol-target overlay. The endpoint derives cashWeight defensively from sum(equity weights) rather than trusting this value directly, but it signals that SGOV data should be included.
   */
  sgovWeight?: number;
}