# lib/api-zod/src/generated/types/portfolioHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Equity Ranking and Risk API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioHistoryResponse {
  /** ISO-8601 date labels — length = daysUsed + 1 (includes base date at index 0) */
  dates: string[];
  /** Cumulative NAV starting at 100 — length = daysUsed + 1 */
  nav: number[];
  /** Drawdown from running peak (%) — length = daysUsed + 1 */
  drawdown: number[];
  /** Total return over the window (%) */
  totalReturn: number;
  /** Maximum drawdown over the window (%) */
  maxDrawdown: number;
  /** Annualized portfolio volatility of log-returns (%) */
  annualizedVol: number;
  /** Number of shared valid daily log-return observations used */
  numDays: number;
  /** Sum of all equity holding weights actually applied in computation */
  investedWeight: number;
  /** Residual cash weight = max(0, 1 - investedWeight) */
  cashWeight: number;
  /** Alias for numDays — number of shared valid observations */
  daysUsed: number;
  /** "sgov" when SGOV prices were available; "zero" fallback otherwise */
  cashMethod: string;
}