# lib/api-zod/src/generated/types/portfolioHoldingRisk.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Equity Ranking and Risk API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioHoldingRisk {
  ticker: string;
  /** Final weight after vol-target overlay */
  weight: number;
  /** Base weight before vol-target overlay (sums to 1) */
  baseWeight: number;
  /** Annualized individual volatility */
  vol: number;
  /** Fraction of total portfolio variance contributed by this name (sums to 1) */
  riskContrib: number;
  /** @nullable */
  cluster?: number | null;
}