# lib/api-zod/src/generated/types/rankingsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Equity Ranking and Risk API
 * OpenAPI spec version: 0.1.0
 */
import type { Stock } from "./stock";
import type { UniverseAudit } from "./universeAudit";

export interface RankingsResponse {
  stocks: Stock[];
  total: number;
  /** @nullable */
  cachedAt?: string | null;
  clusterCount: number;
  audit?: UniverseAudit;
}