# lib/api-zod/src/generated/types/universeAudit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Equity Ranking and Risk API
 * OpenAPI spec version: 0.1.0
 */
import type { UniverseAuditExclusions } from "./universeAuditExclusions";
import type { UniverseAuditSectorBreakdown } from "./universeAuditSectorBreakdown";

export interface UniverseAudit {
  preFilterCount?: number;
  postFilterCount?: number;
  exclusions?: UniverseAuditExclusions;
  sectorBreakdown?: UniverseAuditSectorBreakdown;
  qualityCoverage?: string;
  qualityPct?: number;
  activeFilters?: string[];
}